split integration tests from the rest
